requirements.txt
pyfiglet
langchain
langchain-google-gena
grpcio-tools
